% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercise1.R
\name{checkHeight}
\alias{checkHeight}
\title{calcaulte the difference from the sex-specific or the overall mean height}
\usage{
checkHeight(students.input, sex.specific)
}
\arguments{
\item{students.input}{the student dataframe}

\item{sex.secific}{the student sex}
}
\value{
result.frame
}
\description{
calculate
}
\examples{
checkHeight(students.input = students,sex.specific=FALSE)
checkHeight(students.input = students,sex.specific=TRUE)
}
\seealso{
\code{\link[utils]{head}}
}
