% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean}
\alias{mean}
\alias{meanfunction}
\title{mean value of the vector elements}
\usage{
mean(x)
}
\arguments{
\item{x}{the array input}
}
\value{
the mean value
}
\description{
\code{mean} returns the mean of all the values present in its arguments
}
\details{
This is a generic function: methods can be defined for it directly or via the
\code{\link{mean}} group generic.
}
\examples{
mean(c(1,2,3))
}
