% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stu.R
\docType{data}
\name{students}
\alias{students}
\title{the information of students}
\format{A data frame with 10 rows and 5 variables
\describe{
  \item{name}{name,the first name of students}
  \item{weight}{weight,the weight of students}
  ...
}}
\usage{
students
}
\description{
A dataset containing the name,height and other attributes of students
students
}
\keyword{datasets}
